requirements.txt
langchain
tiktoken
openai
google-generativeai
faiss-cpu
pypdf2
python-dotenv
transformers
ipykernel
flask
flask-login
flask-migrate
flask-mail
flask-cors
sqlalchemy
unstructured
python-telegram-bot
dill